requirements.txt
annotated-types==0.7.0
anyio==4.8.0
av==14.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
filelock==3.17.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
huggingface-hub==0.29.1
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pgvector==0.3.6
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
sentencepiece==0.2.0
setuptools==75.8.2
sniffio==1.3.1
soundfile==0.13.1
speechbrain==1.0.2
SQLAlchemy==2.0.38
starlette==0.46.0
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
